+++ lib/rubocop/cop/layout/indentation_width.rb
@@ -54,6 +54,10 @@ module RuboCop
 
         SPECIAL_MODIFIERS = %w[private protected].freeze
 
+        def_node_matcher :access_modifier?, <<-PATTERN
+          [(send ...) access_modifier?]
+        PATTERN
+
         def on_rescue(node)
           _begin_node, *_rescue_nodes, else_node = *node
           check_indentation(node.loc.else, else_node)
@@ -161,21 +165,35 @@ module RuboCop
         private
 
         def check_members(base, members)
-          check_indentation(base, members.first)
+          check_indentation(base, select_check_member(members.first))
 
           return unless members.any? && members.first.begin_type?
 
           if indentation_consistency_style == 'rails'
-            each_member(members) do |member, previous_modifier|
-              check_indentation(previous_modifier, member,
-                                indentation_consistency_style)
-            end
+            check_members_for_rails_style(members)
           else
             members.first.children.each do |member|
               next if member.send_type? && member.access_modifier?
 
               check_indentation(base, member)
             end
+          end
+        end
+
+        def select_check_member(member)
+          return unless member
+
+          if access_modifier?(member.children.first)
+            member.children.first
+          else
+            member
+          end
+        end
+
+        def check_members_for_rails_style(members)
+          each_member(members) do |member, previous_modifier|
+            check_indentation(previous_modifier, member,
+                              indentation_consistency_style)
           end
         end